+++ checkmarx-ast-eclipse-plugin/src/com/checkmarx/eclipse/startup/PluginStartup.java
@@ -9,7 +9,7 @@ import org.eclipse.ui.PlatformUI;
 import com.checkmarx.eclipse.common.utils.CxLogger;
 import com.checkmarx.eclipse.common.listener.IProjectLifecycleListener;
 import com.checkmarx.eclipse.common.properties.Preferences;
-import com.checkmarx.eclipse.devassist.ui.findings.realtime.CheckmarxEditorListener;
+import com.checkmarx.eclipse.devassist.backend.listener.CheckmarxEditorListener;
 import com.checkmarx.eclipse.devassist.backend.GlobalScannerController;
 import com.checkmarx.eclipse.devassist.backend.listener.ProjectLifecycleListener;
