+++ devassist-lib/META-INF/MANIFEST.MF
@@ -24,5 +24,4 @@ Require-Bundle: com.checkmarx.eclipse.common,
  org.eclipse.e4.ui.css.swt.theme
 Import-Package: org.osgi.service.event;version="1.4.1"
 Export-Package: com.checkmarx.eclipse.devassist.backend,
- com.checkmarx.eclipse.devassist.backend.listener,
- com.checkmarx.eclipse.devassist.ui.findings.realtime
+ com.checkmarx.eclipse.devassist.backend.listener
